requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.10.0
torchvision==0.25.0
Pillow==10.2.0
numpy==1.24.3
gunicorn==21.2.0